+++ example/android/app/src/androidTest/java/com/instabug/instabugflutterexample/InvokeInstabugUITest.java
@@ -26,10 +26,11 @@ public class InvokeInstabugUITest {
     @Test
     public void ensureInstabugInvocati1on() throws InterruptedException {
         disableScreenShotByMediaProjection();
+        Thread.sleep(1000);
         onView(withResourceName("instabug_floating_button")).perform(click());
-        Thread.sleep(5000);
+        Thread.sleep(1000);
         onView(withText("Report a bug")).perform(click());
-        Thread.sleep(5000);
+        Thread.sleep(1000);
         onView(withResourceName("instabug_edit_text_email")).perform(replaceText("inst@bug.com"));
         onView(withResourceName("instabug_bugreporting_send")).perform(click());
         onView(withResourceName("instabug_success_dialog_container")).perform(click());