+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Master
+
+* You can now initialize Instabug on Android by calling `Instabug.start` from Javascript, instead of calling the builder method from the application class.
+* Updates native SDK dependencies to 8.6.2
+
 ## Version 8.6.1 (2019-08-26)
 
 * Bumps version to 8.6 to be in sync with other platforms.